change the server of the precomp bounds and sea to thkukuk.de, use *-latest.zip , search for new files only with option, older local versions can be used with option, list versions. Compile bounds.zip local with a new option for planet_up.py. some changes to simplify the code

+++ planet_up.py
@@ -22,6 +22,7 @@ __email__ = "weigelt.bernd@web.de"
 __status__ = "RC"
 
 import os
+import argparse
 
 WORK_DIR = (os.environ['HOME'] + "/map_build/")
 
@@ -66,18 +67,45 @@ is_there(WORK_DIR, hint)
 
 os.chdir(WORK_DIR)
 
+"""
+argparse
+
+"""
+
+parser = argparse.ArgumentParser(
+        prog='PROG',
+        formatter_class=argparse.RawDescriptionHelpFormatter,
+        description=('''\
+
+            To create and update a local copy of the OSM planet file
+            
+            There is only one option to create new precomp boundaries 
+            from the planet file, useful if the server thkukuk.de is not reachable
+
+        '''))
+        
+        
+parser = argparse.ArgumentParser(
+         prog='PROG',
+         formatter_class=argparse.ArgumentDefaultsHelpFormatter)
+
+parser.add_argument('-nb', '--new_bounds', action="store_true", help="""create new boundaries""")
+
+args = parser.parse_args()
+
+
 for tool in ['osmconvert', 'osmupdate']:
   hint = tool + " missed, please use mk_osmtools to build it from sources"
   checkprg(tool, hint)
 
-if os.path.exists("planet.o5m") == True:
+if os.path.exists("o5m/planet.o5m") == True:
   print()
-  printerror("please move planet.o5m to " + WORK_DIR + "o5m/")
+  printerror("please move planet.o5m to " + WORK_DIR + "planet/")
   quit()
 
-if os.path.exists("o5m/planet.o5m") == False:
+if os.path.exists("planet/planet.o5m") == False:
   print()
-  os.chdir("o5m/")
+  os.chdir("planet/")
   printinfo("Download started. Size ~50 Gigabytes... please wait! ")
   os.system("wget http://ftp5.gwdg.de/pub/misc/openstreetmap/planet.openstreetmap.org/pbf/planet-latest.osm.pbf")
   os.system("osmconvert planet-latest.osm.pbf -o=planet.o5m")
@@ -119,9 +147,9 @@ config.set('planet', 'buildday', (DATE))
 write_config()
 
 print()
-os.system("osmupdate -v --daily --keep-tempfiles o5m/planet.o5m o5m/planet_new.o5m")
+os.system("osmupdate -v --daily --keep-tempfiles planet/planet.o5m planet/planet_new.o5m")
 
-os.chdir("o5m/")
+os.chdir("planet/")
 
 if os.path.exists("planet_new.o5m"):
   os.rename("planet.o5m", "planet_temp.o5m")
@@ -133,41 +161,39 @@ if os.path.exists("planet_new.o5m"):
 create the bounds from planet
 
 """
-option_mkgmap_path = WORK_DIR + config['runtime']['mkgmap'] + "/mkgmap.jar "
+if args.new_bounds:
+  option_mkgmap_path = WORK_DIR + config['runtime']['mkgmap'] + "/mkgmap.jar "
 
-if config['java']['agh'] == "1":
-  option_java_heap = " -XX:+AggressiveHeap "
-else:
-  option_java_heap = " " + config['java']['xmx'] + " " + config['java']['xms'] + " " 
+  if config['java']['agh'] == "1":
+    option_java_heap = " -XX:+AggressiveHeap "
+  else:
+    option_java_heap = " " + config['java']['xmx'] + " " + config['java']['xms'] + " " 
 
-print()
-command_line = ("osmfilter -v planet.o5m  --keep-nodes= --keep-ways-relations='boundary=administrative =postal_code postal_code=' --drop-tags='created_by= source= building= highway= maxspeed= surface= oneway= note= natural= lit=' -o=bounds_data.o5m")
-print()
-os.system(command_line)
+  print()
+  command_line = ("osmfilter -v planet.o5m  --keep-nodes= --keep-ways-relations='boundary=administrative =postal_code postal_code=' --drop-tags='created_by= source= building= highway= maxspeed= surface= oneway= note= natural= lit=' -o=bounds_data.o5m")
+  print()
+  os.system(command_line)
 
 
-command_line = ("java -ea " +
+  command_line = ("java -ea " +
                 option_java_heap +
                 " -cp " +
                 option_mkgmap_path +
                 " uk.me.parabola.mkgmap.reader.osm.boundary.BoundaryPreprocessor " +
                 " bounds_data.o5m bounds ")
-print()
-os.system(command_line)
+  print()
+  os.system(command_line)
 
 
-import shutil
+  import shutil
 
-zipf = "bounds_" + DATE
-shutil.make_archive(zipf, 'zip', "bounds")
-shutil.move("bounds_" + DATE + ".zip", WORK_DIR)
-shutil.rmtree("bounds")
-os.remove("bounds_data.o5m")
+  zipf = "bounds_" + DATE
+  shutil.make_archive(zipf, 'zip', "bounds")
+  shutil.move("bounds_" + DATE + ".zip", WORK_DIR + "precomp")
+  shutil.rmtree("bounds")
+  os.remove("bounds_data.o5m")
 
-os.chdir(WORK_DIR)
-
-config.set('bounds', 'bounds', DATE)
-write_config()
+  os.chdir(WORK_DIR)
 
 printinfo("Habe fertig!")
 quit()
